Adds Reddit Login
Reddit seems to have made some changes lately to the site which has
broken this scripts ability to screenshot reddit comments.
It was identified that when a user is logged in and when a user is not
logged in, the reddit layout is different which is what is causing the
screenshotting to break.
This PR adds in code so that when PlayWright launches the chrome browser
it will perform a Reddit Login first using praw_username and
praw_password that is specified in the config/auth.py file.
Once the browser logs in successfully the existing code is then able to
screenshot the reddit comments again.